Move `TapAuditor` into separate file.
